jlens drift: accumulate in fp64, and retro-correct the recorded cosines
jac_drift divided an fp32 torch.dot by a product of fp32
torch.linalg.vector_norm. dot is accurate to 6e-6; vector_norm is not --
it accumulates d_model^2 terms into one running sum until the smallest
are absorbed, underestimating |J| by 0.69% at 27b. The quotient was
therefore inflated by 1/(r_t*r_0).

The diagnostic is cos(J0,J0), which is exactly 1 by definition and
evaluated to 1.0138 in the old path.

This was not the harmless small-drift curiosity the deck recorded. The
bias is positive at every step and every size, and it grows with
d_model -- 0.01 / 0.02 / 0.26 / 0.47 / 1.52 pp for 270m..27b -- so it
was a cross-size bias sitting inside a cross-size claim. Only its
visibility was confined to near-zero drift, where it pushed cos past 1.

Since dot was already accurate the recorded numbers are recoverable:
cos_true = cos * r(J_t) * r(J_0), and r(J_0)^2 comes either from a
bit-identical first checkpoint (exact -- 1b/270m never saved a base
lens but do have relfro==0 at step 1) or from the saved fp32
.base_lens.pt (4b/12b/27b). On 4b both sources are available and agree
to 0.053pp on a 0.26pp correction. All 660 cos>1 layer-steps go to 0;
the 3 that survive at 1b are 7e-7, fp32 round-off in the stored cosine
the factor is derived from.

relfro is left alone: both its terms are norms, so the bias largely
cancels and ~+0.1% relative survives at 4b.

Raw values are preserved in-file as cos_*_fp32raw and alongside as
metrics.jsonl.fp32raw. Final cos_mean moves 0.9756->0.9755 (270m),
0.9564->0.9563 (1b), 0.9430->0.9411 (4b), 0.9002->0.8970 (12b),
0.8842->0.8737 (27b).

+++ scripts/probe_jlens.py
@@ -145,10 +145,16 @@ def concordance(lens, model, probes, position):
 
 
 def jac_drift(base_jac, lens):
+    # fp64 throughout: J_l has d_model^2 entries (28.9M at 27b), and fp32
+    # torch.linalg.vector_norm accumulates them into one running sum until the
+    # smallest terms are absorbed -- it underestimates |J| by 0.69% at 27b while
+    # torch.dot (blocked accumulation) stays exact. Mixing the two made cos
+    # exceed 1.0 by up to 1.5pp, growing with d_model, so the bias was also
+    # cross-size. Doubling costs 2x transient RAM per layer, not per lens.
     rel, cos = [], []
     for L in lens.source_layers:
-        Jt = lens.jacobians[L].detach().float().cpu().flatten()
-        J0 = base_jac[L].flatten()
+        Jt = lens.jacobians[L].detach().double().cpu().flatten()
+        J0 = base_jac[L].flatten().double()
         n0 = torch.linalg.vector_norm(J0)
         rel.append(float(torch.linalg.vector_norm(Jt - J0) / (n0 + 1e-8)))
         nt = torch.linalg.vector_norm(Jt)
